pthread: pthread: Add __pthread_key_create (for libstdc++)
On glibc <2.33, libstdc++'s thread implementation uses the presense
of this internal symbol to determine whether threads are being used.
In its absence, it silently bypasses various synchronization
functionality such as mutexes under the assumption that the program
only has one thread.

The function isn't actually called in libstdc++, so the body doesn't
matter for this use. On glibc, pthread_key_create is defined as an
alias to this function, so just wrap pthread_key_create to match.

Fixes #376

Signed-off-by: Michael Forney <mforney@mforney.org>